requirements.txt
flask==1.1.4
flask-limiter
tornado
requests
python-dotenv==0.17.0
google-api-python-client
grpcio
grpcio-tools
qrcode
Image